requirements.txt
httpx==0.27.0
flask==3.0.3
pyjwt==2.8.0
pycryptodome==3.20.0